Refactor withdraw() unit testing

+++ include/transaction.h
@@ -12,8 +12,9 @@ typedef struct {
 } Transaction;
 
 int deposit(User *users);
-//double amount_to_deposit();
-double withdraw(User *user);
+// double amount_to_deposit();
+// double withdraw(User *user);
+int withdraw(User *user);
 double amount_to_transfer(User *user);
 int transfer(User *user, double amount);
 // void simulate_7_days(User *users, int user_count);
